# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UserController_index",
        "parameters": [
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "provider",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "limit",
            "in": "query",
            "schema": {
              "maximum": 100,
              "exclusiveMinimum": 0,
              "default": 25,
              "type": "integer"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "offset",
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 9007199254740991,
              "default": 0,
              "type": "integer"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "sort_by",
            "in": "query",
            "schema": {
              "default": "created_at",
              "type": "string"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "sort_dir",
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserIndexResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token requerido, invalido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "403": {
            "description": "El usuario autenticado no tiene el rol requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Indexar usuarios",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UserController_store",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStoreRequestDocDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStoreResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token requerido, invalido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "403": {
            "description": "El usuario autenticado no tiene el rol requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "409": {
            "description": "El email ya esta registrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear usuario",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UserController_show",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserShowResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token requerido, invalido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "403": {
            "description": "El usuario autenticado no tiene el rol requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mostrar usuario",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequestDocDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserUpdateResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token requerido, invalido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "403": {
            "description": "El usuario autenticado no tiene el rol requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "409": {
            "description": "El email ya esta registrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar usuario",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UserController_destroy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDestroyResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token requerido, invalido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "403": {
            "description": "El usuario autenticado no tiene el rol requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar usuario",
        "tags": [
          "users"
        ]
      }
    }
  },
  "info": {
    "title": "Bus Impl API",
    "description": "API de negocio actualizada a 2026",
    "version": "2026.06",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorResponseDocDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "ApiValidationErrorResponseDocDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Campo del request que fallo la validacion."
                },
                "message": {
                  "type": "string",
                  "description": "Mensaje legible del error de validacion."
                }
              },
              "required": [
                "field",
                "message"
              ]
            }
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "UserIndexResponseDocDto_Output": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "count": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "previous": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "maximum": 9007199254740991
                },
                "email": {
                  "type": "string",
                  "maxLength": 160,
                  "format": "email",
                  "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                },
                "full_name": {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 120
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "admin",
                    "operador",
                    "auditor",
                    "soporte"
                  ]
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "active",
                    "inactive",
                    "blocked"
                  ]
                },
                "provider": {
                  "type": "string",
                  "enum": [
                    "local",
                    "google"
                  ]
                },
                "google_sub": {
                  "anyOf": [
                    {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 160
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                }
              },
              "required": [
                "id",
                "email",
                "full_name",
                "role",
                "status",
                "provider",
                "google_sub",
                "created_at",
                "updated_at"
              ],
              "additionalProperties": false
            },
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "success",
          "count",
          "next",
          "previous",
          "results"
        ]
      },
      "UserShowResponseDocDto_Output": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "email": {
                "type": "string",
                "maxLength": 160,
                "format": "email",
                "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
              },
              "full_name": {
                "type": "string",
                "minLength": 2,
                "maxLength": 120
              },
              "role": {
                "type": "string",
                "enum": [
                  "admin",
                  "operador",
                  "auditor",
                  "soporte"
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "active",
                  "inactive",
                  "blocked"
                ]
              },
              "provider": {
                "type": "string",
                "enum": [
                  "local",
                  "google"
                ]
              },
              "google_sub": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              }
            },
            "required": [
              "id",
              "email",
              "full_name",
              "role",
              "status",
              "provider",
              "google_sub",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-self-required": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "success",
          "user"
        ]
      },
      "UserStoreRequestDocDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 160,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "full_name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "role": {
            "default": "operador",
            "type": "string",
            "enum": [
              "admin",
              "operador",
              "auditor",
              "soporte"
            ]
          },
          "status": {
            "default": "active",
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "blocked"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "full_name",
          "password"
        ]
      },
      "UserStoreResponseDocDto_Output": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "email": {
                "type": "string",
                "maxLength": 160,
                "format": "email",
                "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
              },
              "full_name": {
                "type": "string",
                "minLength": 2,
                "maxLength": 120
              },
              "role": {
                "type": "string",
                "enum": [
                  "admin",
                  "operador",
                  "auditor",
                  "soporte"
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "active",
                  "inactive",
                  "blocked"
                ]
              },
              "provider": {
                "type": "string",
                "enum": [
                  "local",
                  "google"
                ]
              },
              "google_sub": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              }
            },
            "required": [
              "id",
              "email",
              "full_name",
              "role",
              "status",
              "provider",
              "google_sub",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-self-required": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "success",
          "user"
        ]
      },
      "UserUpdateRequestDocDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 160,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "full_name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "role": {
            "default": "operador",
            "type": "string",
            "enum": [
              "admin",
              "operador",
              "auditor",
              "soporte"
            ]
          },
          "status": {
            "default": "active",
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "blocked"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128
          }
        }
      },
...
      "UserDestroyResponseDocDto_Output": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "message": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
    }
  }
}